cdk/app.py
#!/usr/bin/env python3
import aws_cdk as cdk
from stacks.pipeline_stack import PipelineStack

app = cdk.App()

PipelineStack(
    app,
    "StocksPipelineStack",
    description="Stocks serverless pipeline — EventBridge, Lambda, DynamoDB, API Gateway, S3",
    env=cdk.Environment(
        account=app.node.try_get_context("account"),
        region=app.node.try_get_context("region") or "us-east-1",
    ),
)

app.synth()
